music-fetcher/requirements.txt
psycopg2-binary
yt-dlp
beets
structlog
python-dotenv
pylast